job #11496 more xfce setup

+++ buildmt/setup.sh
@@ -18,14 +18,23 @@ sudo bash install-package-dependencies.sh
 printf "newpass\nnewpass\n\n" | vnc4passwd
 
 # configure xrdp
-echo "xfce4-session" > /build/buildmt/jenkins-home/.xsession
-echo "xterm*faceName: DejaVu Snas Mono Book" > /build/buildmt/jenkins-home/.Xresources
-echo "xterm*faceSize: 11" >> /build/buildmt/jenkins-home/.Xresources
-xrdb -merge ~/.Xresources
-TMPFILE=`mktemp`
-awk -v found1=0 -v found2=0 '/\[Xvnc\]/{found1=1;}; /^$/{if (found1 && !found2) {found2=1;print "param=-SecurityTypes\nparam=None";}}; //{print $0;}' /etc/xrdp/sesman.ini > $TMPFILE
-sudo cp $TMPFILE /etc/xrdp/sesman.ini
-sudo /etc/init.d/xrdp restart
+cd $DIR
+if [ ! -e ~/.xsession ]; then
+    sudo /etc/init.d/xrdp stop
+    echo "xfce4-session" > ~/.xsession
+    echo "xterm*faceName: DejaVu Snas Mono Book" > ~/.Xresources
+    echo "xterm*faceSize: 11" >> ~/.Xresources
+    xrdb -merge ~/.Xresources
+    TMPFILE=`mktemp`
+    awk -v found1=0 -v found2=0 '/\[Xvnc\]/{found1=1;}; /^$/{if (found1 && !found2) {found2=1;print "param=-SecurityTypes\nparam=None";}}; //{print $0;}' /etc/xrdp/sesman.ini > $TMPFILE
+    sudo cp $TMPFILE /etc/xrdp/sesman.ini
+    xvfb-run bash start-xfce.sh
+    TMPFILE=`mktemp`
+    sed '/switch_window_key/d' ~/.config/xfce4/xfconf/xfce-perchannel-xml/xfce4-keyboard-shortcuts.xml > $TMPFILE
+    cp $TMPFILE ~/.config/xfce4/xfconf/xfce-perchannel-xml/xfce4-keyboard-shortcuts.xml
+    echo $TMPFILE
+    sudo /etc/init.d/xrdp restart
+fi
 
 # install bridgepoint
 cd $DIR
